feat(novel): Add chapter/volume management with version history
- Add novel_chapter crate with Volume, Chapter, ChapterVersion data structures
- Implement per-chapter storage with metadata.json, content.md, and version history
- Add novel_chapters_panel with tree view UI for volumes and chapters
- Support CRUD operations: create/delete/rename chapters and volumes
- Add novel_ai_panel for AI-assisted writing

Storage design:
- Each chapter stored in separate directory: chapters/{id}/
- metadata.json stores chapter metadata
- content.md stores chapter content
- history/v{n}.json stores version snapshots

+++ crates/gpui/src/platform/mac/dispatcher.rs
@@ -87,7 +87,7 @@ impl PlatformDispatcher for MacDispatcher {
 
         unsafe {
             dispatch_async_f(
-                dispatch_get_global_queue(queue_priority, 0),
+                dispatch_get_global_queue(queue_priority as isize, 0),
                 context,
                 Some(trampoline as unsafe extern "C" fn(*mut c_void)),
             );
